Time: 3 ms (28.22%), Space: 8.4 MB (82.61%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -7,7 +7,7 @@ public:
             x = -x;
             sign = 0;
         }
-        long int y = 0;
+        long long int y = 0;
         while(x > 0) {
             y = y*10 + x%10;
             x /= 10;
